% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_server.R
\name{bin.width}
\alias{bin.width}
\title{Fonctions nécessaires pour faire tourner Oser
Création du graphique d'entrée (élicitation)}
\usage{
bin.width(mini, maxi, nBins = 10)
}
\description{
Fonctions nécessaires pour faire tourner Oser
Création du graphique d'entrée (élicitation)
}
