% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_server.R
\name{calc_distrib}
\alias{calc_distrib}
\title{Calcul des distributions par rapport à la sortie de l'elicitation}
\usage{
calc_distrib(myfit, mini, maxi)
}
\description{
Calcul des distributions par rapport à la sortie de l'elicitation
}
