% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_elicit.R
\name{normal.error}
\alias{normal.error}
\title{Fonctions nécessaires pour elicit de SHELF => fonctions non perso}
\usage{
normal.error(parameters, values, probabilities, weights)
}
\description{
Fonctions nécessaires pour elicit de SHELF => fonctions non perso
}
