% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_server.R
\name{plot_elicit}
\alias{plot_elicit}
\title{Fonction pour definir le graphique permettant l'elicitation}
\usage{
plot_elicit(
  name,
  chips,
  mini,
  maxi,
  left,
  right,
  gridHeight = 10,
  fs = 12,
  nBins = 10
)
}
\description{
Fonction pour definir le graphique permettant l'elicitation
}
